chatbot_sctb/requirements.txt
flask
requests
beautifulsoup4
langchain
langchain_openai
pinecone-client
sentence-transformers